[StorePathParts] move to nix package
